refactor: :recycle: rename imports based on change to sub-package name
