+++ src/seedcase_sprout/core/read_properties.py
@@ -1,7 +1,7 @@
 from pathlib import Path
 
+from seedcase_sprout.core.check_datapackage.check_error_matcher import CheckErrorMatcher
 from seedcase_sprout.core.check_is_file import check_is_file
-from seedcase_sprout.core.checks.check_error_matcher import CheckErrorMatcher
 from seedcase_sprout.core.properties import PackageProperties
 from seedcase_sprout.core.read_json import read_json
 from seedcase_sprout.core.sprout_checks.check_properties import check_properties
