+++ src/seedcase_sprout/core/sprout_checks/exclude_non_sprout_resource_errors.py
@@ -1,4 +1,4 @@
-from seedcase_sprout.core.checks.check_error import CheckError
+from seedcase_sprout.core.check_datapackage.check_error import CheckError
 
 
 def exclude_non_sprout_resource_errors(
