+++ src/seedcase_sprout/core/sprout_checks/get_sprout_package_errors.py
@@ -1,4 +1,4 @@
-from seedcase_sprout.core.checks.check_error import CheckError
+from seedcase_sprout.core.check_datapackage.check_error import CheckError
 from seedcase_sprout.core.sprout_checks.check_fields_present import (
     check_fields_present,
 )
