+++ src/seedcase_sprout/core/sprout_checks/required_fields.py
@@ -1,13 +1,13 @@
-from seedcase_sprout.core import checks
-from seedcase_sprout.core.checks.required_fields import (
+from seedcase_sprout.core import check_datapackage
+from seedcase_sprout.core.check_datapackage.required_fields import (
     RequiredFieldType,
 )
 
 # Sprout-specific required fields and their types
 
 PACKAGE_SPROUT_REQUIRED_FIELDS = (
-    checks.PACKAGE_REQUIRED_FIELDS
-    | checks.PACKAGE_RECOMMENDED_FIELDS
+    check_datapackage.PACKAGE_REQUIRED_FIELDS
+    | check_datapackage.PACKAGE_RECOMMENDED_FIELDS
     | {
         "title": RequiredFieldType.str,
         "description": RequiredFieldType.str,
@@ -17,7 +17,7 @@ PACKAGE_SPROUT_REQUIRED_FIELDS = (
 )
 PACKAGE_SPROUT_REQUIRED_FIELDS.pop("resources", None)
 
-RESOURCE_SPROUT_REQUIRED_FIELDS = checks.RESOURCE_REQUIRED_FIELDS | {
+RESOURCE_SPROUT_REQUIRED_FIELDS = check_datapackage.RESOURCE_REQUIRED_FIELDS | {
     "title": RequiredFieldType.str,
     "description": RequiredFieldType.str,
 }
