+++ src/seedcase_sprout/core/update_package_properties.py
@@ -1,4 +1,4 @@
-from seedcase_sprout.core.checks.check_error_matcher import CheckErrorMatcher
+from seedcase_sprout.core.check_datapackage.check_error_matcher import CheckErrorMatcher
 from seedcase_sprout.core.properties import PackageProperties
 from seedcase_sprout.core.sprout_checks.check_package_properties import (
     check_package_properties,
