+++ src/seedcase_sprout/core/write_package_properties.py
@@ -1,6 +1,6 @@
 from pathlib import Path
 
-from seedcase_sprout.core.checks.check_error_matcher import CheckErrorMatcher
+from seedcase_sprout.core.check_datapackage.check_error_matcher import CheckErrorMatcher
 from seedcase_sprout.core.properties import PackageProperties
 from seedcase_sprout.core.sprout_checks.check_properties import check_properties
 from seedcase_sprout.core.write_json import write_json
