+++ tests/core/sprout_checks/test_check_fields_present.py
@@ -1,6 +1,6 @@
 from pytest import mark
 
-from seedcase_sprout.core.checks.required_fields import RequiredFieldType
+from seedcase_sprout.core.check_datapackage.required_fields import RequiredFieldType
 from seedcase_sprout.core.sprout_checks.check_fields_present import (
     check_fields_present,
 )
