+++ tests/core/sprout_checks/test_check_package_properties_sprout.py
@@ -1,6 +1,6 @@
 from pytest import fixture, mark, raises
 
-from seedcase_sprout.core.checks.check_error_matcher import CheckErrorMatcher
+from seedcase_sprout.core.check_datapackage.check_error_matcher import CheckErrorMatcher
 from seedcase_sprout.core.properties import PackageProperties
 from seedcase_sprout.core.sprout_checks.check_package_properties import (
     check_package_properties,
